libs/start.py
"""
author: Shiv, [please add your name if you update the script]
email: shivkj001@gmail.com
date: 20-02-2021
"""
import datetime as dt
import json
import math
import re
from collections import Counter, defaultdict, deque
from dataclasses import dataclass, field
from functools import lru_cache, partial, wraps
from itertools import combinations, product
from typing import Dict, List, Tuple

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
from dateutil.parser import parse
from tqdm import tqdm
