libs/requirements.txt
dask
isort
matplotlib
nltk
numpy
pandas
Pillow
pytest
python-dateutil
requests
scikit-learn
scipy
seaborn
Shapely
tqdm
xlrd